// src/enums/routes.enum.ts
export enum RoutesApp {
  LOGIN = 'login',
  DASHBOARD = 'dashboard', //dashboard TEST
  SETTINGS = 'settings',
  CATALGOS = 'catalogs',
  CREATE_CATALOGS = 'create-catalog',
  PROGRAMS = 'programs',
  SUBJECT_PROGRAMS = 'subject-programs',
  CREATE_UPDATE_SUBJECT_GUIDE = 'create-update-guide',
  LEARNING_EVALUACION = 'learning-evaluation',
  NEWS = 'news',
  CREATE_UPDATE_NEW = 'create-update-new',
  VIEW_NEW = 'view-new',
  CALENDAR = 'calendar',
  LOGOUT = 'logout',
  CREATE_PROGRAM = 'create-program',
  EDIT_PROGRAM = 'edit-program',
  CURRICULUM = 'curriculum',
  CURRICULAR_CYCLE = 'curricular-cycle',
  CURRICULAR_MONITORING = 'curricular-monitoring',
  SYLLABUS = 'syllabus',
  RAE = 'rae',
  SPECIFIC_SUMARY = 'specific-summary',
  CURRICULAR_COMPONENT = 'currucular-component',
  MODULE_SYLLABUS = 'module-syllabus',
  CURRICULAR_OUTPUT = 'sub-core',
  ACADEMIC_CREDITS = 'academic-credit',
  CORE_AND_SUBCORES = 'core-subcore',
  TECHNICAL_TECHNOLOGICAL_PROGRAMS = 'technical-tecnological',
  UPDATE_AUTHORIZATIONS = 'update-authorization',
  PROBLEM_BANK = 'problem_bank',
  STUDY_PLAN = 'study_plan',
  GUIDELINES = 'guidelines',
  COMMITTES = 'committes',
  MY_ACADEMIC_PROGRAM = 'my_academic_program',
  INFORMATION_INTEGRATING_CORE = 'information_integrating_core',
  NEW_INFORMATION_INTEGRATING_CORE = 'new',
  VIEW_INFORMATION_INTEGRATING_CORE = 'view',
  VIEW_INF_INT_CORE_EDIT = 'edit_subcore',
  SUMMARY = 'summary',
  TRAINING = 'training',
  DIRECTORY = 'directory',
  FACULTY = 'faculty',
  SECONG_LENGUAGE_LEARNINNG = 'second-language-learning',
  SEARCH = 'search',
  ROLES = 'roles',
  EPORTAFOLIO = 'eportafolio',
  ACADEMIC_RECORD = 'academic_record',
  MODULES = 'modules',
  PERMISSION_MODULE = 'permission_module',
  DOC_GUIDELINES = 'doc-guidelines',
  DOC_INSTRUMENTS = 'doc-instruments',
  TEMPLATE = 'template',
  FLOWS = 'flow',
  UPDATE_TEMPLATE = 'update-template',
  STEP_FLOW = 'step',
}